Split the environment componenet to Weather.componenet and Clock.component
Added functionality to clock and weather for coniditional rendering.
