+++ package.json
@@ -41,11 +41,13 @@
     "last 2 versions"
   ],
   "dependencies": {
+    "@types/classnames": "^2.2.10",
     "@types/moment": "^2.13.0",
     "@types/react-router-dom": "^5.1.5",
     "app-root-path": "^3.0.0",
     "bootstrap": "^4.4.1",
-    "dayjs": "^1.9.3",
+    "classnames": "^2.2.6",
+    "dayjs": "^1.9.4",
     "file-system": "^2.2.2",
     "moment": "^2.28.0",
     "react": "^16.13.1",
