+++ src/styles/components/weather.scss
@@ -1,3 +1,18 @@
 .weather {
   color: $primary-text-color;
+
+  &__text {
+    font-weight: bold;
+  }
+  &__info {
+    flex-direction: column;
+    justify-content: space-evenly;
+    display: flex;
+  }
+  &__cold {
+    color: aqua;
+  }
+  &__hot {
+    color: crimson;
+  }
 }
